Started with state management. Tried to keep the state at the App level. # Please enter the commit message for your changes. Lines starting # with '#' will be ignored, and an empty message aborts the commit. # # On branch master # Changes to be committed: #	modified:   src/App.js     Removed unnecessary imports and created the bulk of JSX and state management here. #	new file:   src/components/Checkbox.js     Simple component that is just displaying the contents of the checkbox. This can definitely be cleaned up. #	modified:   src/index.js     Removed 'reportWebVitals'. #	deleted:    src/logo.svg     Deleted the default React logo. #	deleted:    src/reportWebVitals.js     Deleted reportWebVitals because I was not using it. #
There is a lot of clean up to do. As much as possible should be decoupled from App.js because it's already becoming a mess. App.js is going to be mostly for state management since it's at the highest level (in this case) as to avoid prop drilling. All states should flow down from App.js.